refactor(partidas-de-futebol-mensageria): ajusta retorno para ResponseEntity com status 202 e renomeia nomenclaturas de pasta de config e atributo.
- Atualiza CriarClubeController para retornar ResponseEntity com status HTTP 202 (Accepted) ao criar clube
- Altera nomenclaturas nos arquivos de configuração para manter consistência e clareza
